aws_handler.py
import boto3
import os
from dotenv import load_dotenv

load_dotenv()

def get_s3_client():
    return boto3.client(
        "s3",
        aws_access_key_id=os.getenv("AWS_ACCESS_KEY"),
        aws_secret_access_key=os.getenv("AWS_SECRET_KEY"),
        region_name=os.getenv("AWS_REGION")
    )

def upload_stream_to_s3(s3_client, bucket_name, key, stream):
    s3_client.upload_fileobj(stream, bucket_name, key)
