requirements.txt
boto3
streamlit
requests
msal
python-dotenv
